+++ CHANGELOG.md
@@ -9,5 +9,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Add missing server_run analytics event
+- Run analytics only on version specified in config
 
 ### Misc